REFACTOR: Imports and header component #7
