+++ src/components/index.ts
@@ -1,10 +1,11 @@
 import GraphContainer from "./graph/GraphContainer.vue";
 import PainelControl from "./painel/PainelControl.vue";
-import DesktopHeader from './DefaultHeader/DesktopHeader.vue';
-import MobileHeader from './DefaultHeader/MobileHeader.vue';
-import ListTitles from './DefaultHeader/ListTitles/ListTitles.vue';
-import TextLogo from './DefaultHeader/TextLogo/TextLogo.vue';
-import LoginButton from './DefaultHeader/LoginButton/LoginButton.vue';
+import DesktopHeader from './header/DesktopHeader.vue';
+import MobileHeader from './header/MobileHeader.vue';
+import ListTitles from './header/titles/ListTitles.vue';
+import TextLogo from './header/logo/TextLogo.vue';
+import LoginButton from './header/buttons/LoginButton.vue';
+import DefaultHeader from './header/DefaultHeader.vue'
 
 export {
     GraphContainer,
@@ -13,5 +14,6 @@ export {
     MobileHeader,
     ListTitles,
     TextLogo,
-    LoginButton
+    LoginButton,
+    DefaultHeader
 };
