+++ src/pages/AboutView.vue
@@ -1,12 +1,8 @@
 <script setup lang="ts">
-
+import { DefaultHeader } from '@/components';
 </script>
 
 <template>
+  <DefaultHeader/>
   <h1>About</h1>
 </template>
-
-
-<style scoped>
-
-</style>
