+++ src/pages/AuthorView.vue
@@ -1,12 +1,8 @@
 <script setup lang="ts">
-
+import { DefaultHeader } from '@/components';
 </script>
 
 <template>
+  <DefaultHeader/>
   <h1>Author</h1>
 </template>
-
-
-<style scoped>
-
-</style>
